Consolidate data preparation and flatten file structure
