+++ .gitignore
@@ -1,3 +1,6 @@
+# _extensions
+*_extensions*
+
 # .clocignore
 .clocignore
