+++ bibat/{{cookiecutter.repo_name}}/Makefile
@@ -1,7 +1,8 @@
-.PHONY: clean-runs clean-plots clean-stan clean-all analysis
+.PHONY: clean-runs clean-plots clean-stan clean-all analysis env
 
 ACTIVATE_VENV = .venv/bin/activate
 REQUIREMENTS_FILE = requirements.txt
+SRC = {{cookiecutter.repo_name}}
 
 ifeq ($(OS),Windows_NT)
 	INSTALL_CMDSTAN_FLAGS = --compiler
@@ -16,29 +17,29 @@ $(ACTIVATE_VENV):
 env: $(ACTIVATE_VENV) $(REQUIREMENTS_FILE) $(CMDSTAN)
 	. $(ACTIVATE_VENV) && (\
 	  python -m pip install --upgrade pip; \
-	  python -m pip install -r $(REQUIREMENTS_FILE) --quiet; \
+		python -m pip install -e .; \
 	  install_cmdstan $(INSTALL_CMDSTAN_FLAGS); \
 	)
 
 analysis: env
 	. $(ACTIVATE_VENV) && (\
 	  {% if cookiecutter.create_tests_directory %}python -m pytest || exit 1; \{% endif %}
-	  python prepare_data.py || exit 1; \
-	  python sample.py || exit 1; \
-	  jupyter execute investigate.ipynb || exit 1; \
+	  python $(SRC)/prepare_data.py || exit 1; \
+	  python $(SRC)/sample.py || exit 1; \
+	  jupyter execute $(SRC)/investigate.ipynb || exit 1; \
 	)
 
 clean-stan:
-	$(RM) $(shell find ./src/stan -perm +100 -type f) # remove binary files
-	$(RM) ./src/stan/*.hpp
+	$(RM) $(shell find ./$(SRC)/stan -perm +100 -type f) # remove binary files
+	$(RM) $(SRC)/stan/*.hpp
 
-clean-runs:
-	$(RM) -r results/runs/*/
+clean-inferences:
+	$(RM) $(shell find ./inferences/* -type f -not -name "*.toml")
 
 clean-plots:
-	$(RM) -r results/plots/*.png
+	$(RM) -r plots/*.png
 
 clean-prepared-data:
 	$(RM) -r data/prepared/*/
 
-clean-all: clean-prepared-data clean-stan clean-runs clean-plots
+clean-all: clean-prepared-data clean-stan clean-inferences clean-plots
