+++ bibat/{{cookiecutter.repo_name}}/{{cookiecutter.repo_name}}/__init__.py
