+++ bibat/{{cookiecutter.repo_name}}/{{cookiecutter.repo_name}}/stan/custom_functions.stan
