+++ bibat/{{cookiecutter.repo_name}}/{{cookiecutter.repo_name}}/stan/multilevel-linear-regression.stan
