+++ bibat/{{cookiecutter.repo_name}}/{{cookiecutter.repo_name}}/stan/readme.md
